add: 座標情報のデモデータを追加

+++ docker/postgres/init/4-create_walking_sampels_model_table.sql
@@ -2,19 +2,19 @@
 
 CREATE TABLE particles (
     id VARCHAR(26) PRIMARY KEY,
-    x DECIMAL NOT NULL,
-    y DECIMAL NOT NULL,
-    weight DECIMAL NOT NULL,
-    direction DECIMAL NOT NULL,
+    x INTEGER NOT NULL,
+    y INTEGER NOT NULL,
+    weight INTEGER NOT NULL,
+    direction INTEGER NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
     walking_sample_id VARCHAR(26) REFERENCES walking_samples(id)
 );
 
 CREATE TABLE estimated_positions (
     id VARCHAR(26) PRIMARY KEY,
-    x DECIMAL NOT NULL,
-    y DECIMAL NOT NULL,
-    direction DECIMAL NOT NULL,
+    x INTEGER NOT NULL,
+    y INTEGER NOT NULL,
+    direction INTEGER NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
     walking_sample_id VARCHAR(26) UNIQUE REFERENCES walking_samples(id)
 );
